requirements.txt
# Core Data Science and Analysis
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1
scikit-learn==1.3.0
statsmodels==0.14.0

# Machine Learning Frameworks
xgboost==1.7.6
lightgbm==4.0.0
catboost==1.2

# Deep Learning
torch==2.0.1
tensorflow==2.13.0
keras==2.13.1
torch-geometric==2.3.1

# Bayesian Modeling
pymc==5.6.1
arviz==0.16.1
bambi==0.13.0

# Time Series Analysis
prophet==1.1.4
pmdarima==2.0.3
tslearn==0.6.1

# Reinforcement Learning
gym==0.26.2
stable-baselines3==2.0.0

# Hyperparameter Optimization
optuna==3.2.0
hyperopt==0.2.7

# Data Collection and APIs
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.11.2
aiohttp==3.8.5

# Database
sqlalchemy==2.0.19
psycopg2-binary==2.9.6
redis==4.6.0
pymongo==4.4.1

# Web Framework and API
fastapi==0.100.0
uvicorn[standard]==0.23.1
pydantic==2.1.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.8

# Data Validation and Quality
great-expectations==0.17.12
pandera==0.15.2

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
dash==2.11.1

# Model Interpretability
shap==0.42.1
lime==0.2.0.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
joblib==1.3.1
tqdm==4.65.0
rich==13.5.0

# Data Processing
pyarrow==12.0.1
polars==0.18.7
dask==2023.7.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-asyncio==0.21.1
hypothesis==6.82.0

# Code Quality
black==23.7.0
flake8==6.0.0
mypy==1.4.1
pylint==2.17.4

# Monitoring and Logging
mlflow==2.5.0
wandb==0.15.7
loguru==0.7.0

# Development Tools
jupyter==1.0.0
jupyterlab==4.0.3
ipywidgets==8.0.7
nbconvert==7.7.2

# Caching
diskcache==5.6.1
cachetools==5.3.1

# Task Queue
celery==5.3.1
redis==4.6.0

# Date/Time Handling
python-dateutil==2.8.2
pytz==2023.3

# Web Scraping Tools
lxml==4.9.3
html5lib==1.1

# Additional ML Tools
imbalanced-learn==0.11.0
yellowbrick==1.5

# Graph Analysis (for GNN)
networkx==3.1
dgl==1.1.1

# Natural Language Processing (for text analysis)
nltk==3.8.1
spacy==3.6.0

# Configuration Management
hydra-core==1.3.2
omegaconf==2.3.0

# API Documentation
fastapi-users==12.0.0
fastapi-pagination==0.12.6

# Security
cryptography==41.0.2
bcrypt==4.0.1

# Performance Profiling
memory-profiler==0.61.0
line-profiler==4.0.3

# CORS Support
fastapi-cors==0.0.6

# Environment Variables Validation
pydantic-settings==2.0.2